% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_flights_db.R
\name{build_flights_db}
\alias{build_flights_db}
\title{Build or update the flights database}
\usage{
build_flights_db(site, refresh = FALSE, really = FALSE)
}
\arguments{
\item{site}{Site abbreviation}

\item{refresh}{Recreated database from scratch. \strong{Warning:} this will
destroy your existing database, including all assigned scores and
comments. Requires also supplying \code{really = TRUE}.}

\item{really}{If TRUE, allows refresh to recreate the database}
}
\value{
A list of
\item{db}{Site database table}
\item{db_name}{Path and name of database table}
}
\description{
Build or update the flights database for a site, normally called by
\code{screen}; call if you're unable or unwilling to run \code{screen}.
}
\details{
Reads any existing flights_\if{html}{\out{<site>}}.txt from flights directory for site,
builds it or updates it for new or deleted files, saves the new version,
and returns the path/name and table. Finds classes from \code{pars.yml} as
case-insensitive underscore-separated words (after applying name fixes).

Files with changed timestamps are presumed to have been re-downloaded
with gather (as stamps are set in processing). Files shouldn't be
re-downloaded and replaced unless they've changed on the source, so
these files have presumable been repaired. They are refreshed in the
flights database, ready for re-screening.
}
\keyword{internal}
