% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_screen_db.R
\name{build_screen_db}
\alias{build_screen_db}
\title{Read and update or build a screening database for a site}
\usage{
build_screen_db(site)
}
\arguments{
\item{site}{Site abbreviation}
}
\value{
A list of
\item{db}{Site database table}
\item{db_name}{Path and name of database table}
}
\description{
Reads any existing screen_\if{html}{\out{<site>}}.txt from flights directory for site,
builds it or updates it for new or deleted files, saves the new version,
and returns the path/name and table.
}
\keyword{internal}
