% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_python_env.R
\name{create_py_env}
\alias{create_py_env}
\title{Create Python 3.10 env for U-Net}
\usage{
create_py_env()
}
\description{
Rerun this when
\itemize{
\item new user or system
\item Python version changes
\item the environment gets corrupted
}

To install additional Python packages, use \code{virtualenv_install()}
}
