% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuda_check.R
\name{cuda_check}
\alias{cuda_check}
\title{Make sure CUDA is available for GPU jobs}
\usage{
cuda_check(requirecuda, test = TRUE)
}
\arguments{
\item{requirecuda}{If TRUE, job expects a GPU, so do the test. If FALSE, just return.}

\item{test}{If TRUE, does a print test to try to get a better failure error message}
}
\description{
Make sure CUDA is available for GPU jobs
}
\keyword{internal}
