% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_purge.R
\name{db_purge}
\alias{db_purge}
\title{Purge selected fits or maps from database}
\usage{
db_purge(which, db_name, id_name, rows, failed, undo)
}
\arguments{
\item{which}{Which database? Either \code{fit} or \code{map}}

\item{db_name}{Either 'fdb' or 'mdb'}

\item{id_name}{Either \code{fitid} or \code{mapid}}

\item{rows}{Selected rows in the database. Use one of
\itemize{
\item an empty string doesn't purge any rows, but does purge stray fit sidecar files
\item a vector of ids to purge those rows
\item a named list to filter rows. List items are \verb{<field in database> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{failed}{If TRUE, all rows where \code{success} is \code{FALSE} or \code{NA} are purged.
This is an alternative to specifying \code{rows}.}

\item{undo}{Undo previous purges. There is no time limit on undoing, and it doesn't
matter whether you've run more fits or maps since a purge. You may supply either:
\itemize{
\item \code{undo = 'last'}, reverses the previous purge call, and the
database and associated files are restored.
\item a vector of fit ids corresponding to previously purged rows.
Note that you may view purged fits with \code{fitinfo} or \code{mapinfo} with \code{purged = TRUE}.
}}
}
\description{
Called by \code{fitpurge} or \code{mappurge} to purge database rows or restore them, as well
as purging or restoring sidecar files.
}
\keyword{internal}
