% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{Create derived variables such as NDVI and NDRE}
\usage{
derive(
  site,
  pattern1 = "",
  pattern2 = NULL,
  metrics = c("NDVI", "NDWI", "NDRE"),
  window = 3,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. If running in batch mode, each
named site will be run in a separate job.}

\item{pattern1}{Regex filtering rasters, case-insensitive. Default = "" (match all). Note: only
files ending in \code{.tif} are included in any case. \code{pattern1} may alternatively be a vector
of layer names (or patterns if you want to get tricky). See details.}

\item{pattern2}{A second pattern or vector of layer names, used for bivariate metrics. See details.}

\item{metrics}{A list of metrics to apply. Univariate metrics include:
\describe{
\item{NDVI}{Normalized difference vegetation index, (NIR - red) / (NIR + red), an index of biomass}
\item{NDWIg}{Normalized difference water index (green), (green - NIR) / (green + NIR),
primarily helps distingish waterbodies}
\item{NDRE}{Normalized difference red edge index, (NIR - RE) / (NIR + RE), an index of the
amount of chlorophil in a plant}
\item{NDVI_mean}{mean of NDVI in a window, size defined by \code{window}}
\item{NDVI_std}{standard deviation of NDVI in a window, size defined by \code{window}}
Bivariate metrics include:
\item{NDWIswir}{Normalized difference water index (SWIR), (NIR - SWIR) / (NIR + SWIR),
an index of water content in leaves; requires a Mica layer for \code{pattern1}, and a matched
SWIR layer for \code{pattern2}}
\item{delta}{The difference between \code{pattern1} and \code{pattern2}, may be useful for taking a
difference between late-season and early-season DEMs to represent vegetation canopy height}
}}

\item{window}{Window size for NDVI_mean and NDVI_std, in cells; windows are square, so just specify
a single number. Bonus points if you remember to make it odd.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error handling. Use this
for debugging. If you get unrecovered errors, the job won't be added to the jobs database. Has
no effect if local = FALSE.}

\item{comment}{Optional slurmcollie comment}
}
\description{
This function creates any of a number of univariate and bivariate metrics derived from
raster data, such as NDVI and NDRE. Results are written as rasters to the \code{flights} directory,
with the metric included in the name. These metrics will be treated like source layers in
subsequent processing and modeling.
}
\details{
For univariate metrics, supply one or more layer names via \code{pattern1}. All metrics will be
calculated for each layer specified by \code{pattern1}. Results will be named \verb{<layer>__<metric>}.

For bivariate metrics, speicfy matched pairs of layers with \code{pattern1} and \code{pattern2}. It's
best to specify complete names (you can use vectors for each) so the layers are paired properly.
If you're crazy enough to use patterns here, scrutinize the result names carefully.
Results will be named \verb{<layer1>_<layer2>__metric}. At the moment, \code{delta} is the only bivariate
metric.

This fits in the workflow after \code{gather} and before \code{sample}.
}
