% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{describe_channels}
\alias{describe_channels}
\title{Describe channels from image filenames}
\usage{
describe_channels(images)
}
\arguments{
\item{images}{Character vector of image filenames}
}
\value{
Character string like '26 channels: 2023 summer-mid (mica,DEM,NDVI,NDRE,NDWIg), post-mid (mica,NDVI,NDRE)'
}
\description{
Describe channels from image filenames
}
\keyword{internal}
