% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fit.R
\name{do_fit}
\alias{do_fit}
\title{Fit models}
\usage{
do_fit(
  fitid,
  sites,
  name,
  method = "rf",
  vars = NULL,
  exclude = NULL,
  years = NULL,
  maxmissing = 0.05,
  top_importance = 20,
  holdout = 0.2,
  auc = TRUE,
  hyper,
  rep = NULL
)
}
\arguments{
\item{fitid}{Fit id in the fits database}

\item{sites}{Data frame with \code{site} (3 letter code), \code{site_name} (long name), and
\code{datafile} (resolved path and filename of datafile). Sites, paths, and filenames
are vetted by fit - there's no checking here.}

\item{name}{Optional model name}

\item{method}{One of \code{rf} for Random Forest, \code{boost} for AdaBoost. Default = \code{rf}.}

\item{vars}{An optional vector of variables to restrict analysis to. Default = NULL,
all variables. You may use portable names (in full or via regex), file names (in
full or via regex), or search names. If you use file names, with multiple sites,
these must match portable names in any site, which will be used for all sites.}

\item{exclude}{An optional vector of variables to exclude. Names are specified as for
\code{vars}.}

\item{years}{An optional vector of years to restrict variables to.}

\item{maxmissing}{Maximum proportion of missing training points allowed before a
variable is dropped.}

\item{top_importance}{Number of variables to keep for variable importance}

\item{holdout}{Proportion of points to hold out. For Random Forest, this specifies
the size of the single validation set, while for boosting, it is the size of each
of the testing and validation sets.}

\item{auc}{If TRUE, calculate class probabilities so we can calculate AUC.}

\item{hyper}{Hyperparameters. \emph{\strong{To be defined.}}}

\item{rep}{Throwaway argument for \code{slurmcollie}.}
}
\description{
Fit models
}
