% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_map.R
\name{do_map}
\alias{do_map}
\title{Produce geoTIFF maps of predicted vegetation cover from fitted models}
\usage{
do_map(site, fitid, fitfile, clip, result, rep = NULL)
}
\arguments{
\item{site}{Three letter site code. If fitting from a fit id that was built
on a single site, you may omit \code{site} to map the same site (this is the
most common situation). If you want to map sites other than the site the
model was built on, or the model was built on mutiple sites, \code{site} is
required.}

\item{fitid}{id in the fits database (NULL if not specified)}

\item{fitfile}{Full specification of an RDS file with the fit object}

\item{clip}{Optional clip, vector of \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}}

\item{result}{Optional result name. Default is
\verb{map_<site>_<fit id>_[clip_<size>_ha]}; if a result name is specified,
the result will be \verb{map_<result>_<site>_<fit id>_[clip_<size>_ha]},
retaining the site and fit id, as omiting these breaks your ability to
track maps back to the fits they're based on.}

\item{rep}{Throwaway argument to make \code{slurmcollie} happy}
}
\description{
Console command to launch a prediction run via \code{do_map}, typically in a batch job on Unity.
}
\details{
Side effects:
\enumerate{
\item writes a geoTIFF, \verb{<result>.tif} with, and a run info file
\item \verb{<runinfo>.RDS}, with the following:
\enumerate{
\item Time taken for the run (s)
\item Maximum memory used (GB)
\item Raster size (M pixel)
\item R error, or NULL for success
}
}

Requires \code{rasterPrep}. Install it with:
\code{remotes::install_github('ethanplunkett/rasterPrep')}
}
