% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sample.R
\name{do_sample}
\alias{do_sample}
\title{Sample Y and X variables for a site}
\usage{
do_sample(
  site,
  pattern,
  n,
  p,
  d,
  classes,
  minscore,
  maxmissing,
  reclass,
  balance,
  balance_excl,
  result,
  transects,
  drop_corr,
  reuse
)
}
\arguments{
\item{site}{site, using 3 letter abbreviation}

\item{pattern}{File names, portable names, regex matching either, or search names
selecting files to sample. See Image naming in
\href{https://github.com/UMass-UAS-Salt-Marsh/marshmap/blob/main/README.md}{README}
for details.W}

\item{n}{Number of total samples to return (up to number available).}

\item{p}{Proportion of total samples to return. Use p = 1 to sample all.}

\item{d}{Mean distance in cells between samples. No minimum spacing is guaranteed.}

\item{classes}{Class or vector of classes in transects to sample. Default is all
classes.}

\item{minscore}{Minimum score for orthos. Files with a minimum score of less than
this are excluded from results. Default is 0, but rejected orthos are always
excluded.}

\item{maxmissing}{Maximum percent missing in orthos. Files with percent missing greater
than this are excluded.}

\item{reclass}{Vector of paired classes to reclassify, e.g., \code{reclass = c(13, 2, 3, 4)}
would reclassify all 13s to 2 and 4s to 3, lumping each pair of classes.}

\item{balance}{If TRUE, balance number of samples for each class. Points will be randomly
selected to match the sparsest class.}

\item{balance_excl}{Vector of classes to exclude when determining sample size when
balancing. Include classes with low samples we don't care much about.}

\item{result}{Name of result file. If not specified, file will be constructed from
site, number of X vars, and strategy.}

\item{transects}{Name of transects file; default is \code{transects}.}

\item{drop_corr}{Drop one of any pair of variables with correlation more than \code{drop_corr}.}

\item{reuse}{Reuse the named file (ending in \verb{_all.txt}) from previous run, rather
than resampling. Saves a whole lot of time if you're changing \code{n}, \code{p}, \code{d}, \code{balance},
\code{balance_excl}, or \code{drop_corr}. ** Not implemented yet! **}
}
\value{
Sampled data table (invisibly)
}
\description{
Sample imagery for a site and create rectangular data files for model
fitting. This is an internal function, called by sample.
}
\details{
\strong{Memory requirements: I've measured up to 28.5 GB.}
}
\keyword{internal}
