% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_area.R
\name{extent_area}
\alias{extent_area}
\title{Give the area of a clip vector}
\usage{
extent_area(clip, units = "ha", crs)
}
\arguments{
\item{clip}{Vector of \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}}

\item{units}{Area units to return}

\item{crs}{Coordinate reference system of clip. Pass the CRS from
the relevant raster (e.g., \code{crs(my_raster)}) to ensure correctness
regardless of the project's coordinate system.}
}
\description{
Give the area of a clip vector
}
\keyword{internal}
