% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldinfo.R
\name{field_from_raster}
\alias{field_from_raster}
\title{Field data for a site from its rasterized transects}
\usage{
field_from_raster(site)
}
\arguments{
\item{site}{Three letter site code}
}
\value{
Data frame with columns \code{year}, \code{poly}, \code{subclass}, and \code{cells}
}
\description{
This is the original approach, counting actual cells in \code{transects.tif}. See \code{fieldinfo}.
}
\keyword{internal}
