% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_db.R
\name{filter_db}
\alias{filter_db}
\title{Filter fits database}
\usage{
filter_db(filter, database)
}
\arguments{
\item{filter}{Specify fits with one of:
\itemize{
\item a vector of \code{ids}
\item 'all' for all fits
\item a named list to filter fits. List items are \verb{<field in fits database> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{database}{Name of database, either \code{fdb} or \code{mdb}}
}
\value{
A vector of rows numbers in the selected database
}
\description{
Filter fits database
}
\keyword{internal}
