% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fits.R
\name{filter_fits}
\alias{filter_fits}
\title{Filter fits database}
\usage{
filter_fits(filter)
}
\arguments{
\item{filter}{Specify fits with one of:
\itemize{
\item a vector of \code{ids}
\item 'all' for all fits
\item a named list to filter fits. List items are \verb{<field in fits database> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}
}
\value{
A vector of rows numbers in \code{the$fdb}
}
\description{
Filter fits database
}
