% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthos.R
\name{find_orthos}
\alias{find_orthos}
\title{Find orthophotos for a site}
\usage{
find_orthos(site, descrip)
}
\arguments{
\item{site}{Site name}

\item{descrip}{Character string with one or more of any of the following,
separated by \code{+}:
\item{file name}{a complete file name (\code{.tif} is optional)}
\item{portable name}{a portable name}
\item{regex}{a regular expression enclosed in \code{{}}, to be
applied to both file names and portable names. Regular expressions
are case-insensitive.}
\item{search name}{a search string (see \code{search_names} for details)}}
}
\value{
Data frame with
\item{row}{row numbers in \verb{flights<site>.txt}}
\item{file}{file names}
\item{portable}{portable names}
}
\description{
Finds orthophotos at a given site from complete file names, portable names,
search names, or regular expressions that match file names or portable names.
These may be mixed and matched, separated by \code{+}.
}
\details{
Note that portable names will be filtered so there is only one result for
each unique portable name. When there are duplicate portable names at a site,
filtering picks the portable name with the highest score, breaking ties by
picking the earliest day in the season. Filtering is only applied to portable
names (perhaps matched by a regex). File names and search names that give
multiple matches are not filtered.
}
