% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finish_map.R
\name{finish_map}
\alias{finish_map}
\title{Finish do_map run, updating maps database}
\usage{
finish_map(jobid, status)
}
\arguments{
\item{jobid}{slurmcollie job id}

\item{status}{Job status}
}
\description{
Finish do_map run, updating maps database
}
