% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_finish.R
\name{fit_finish}
\alias{fit_finish}
\title{Finish fit run}
\usage{
fit_finish(jobid, status)
}
\arguments{
\item{jobid}{Job id to finish for}

\item{status}{Job status}
}
\description{
Finish a fit run:
\itemize{
\item Populate fits database with \code{slurmcollie} stats
\item and with info from \verb{zz_<id>_fit.RDS}, written by \code{do_fit}
\item Copy the log file to the models directory
}
}
