% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitinfo.R
\name{fitinfo}
\alias{fitinfo}
\title{Display information on model fits}
\usage{
fitinfo(
  rows = "all",
  cols = "normal",
  report = NULL,
  sort = "id",
  decreasing = FALSE,
  nrows = NA,
  quiet = FALSE,
  timezone = "America/New_York"
)
}
\arguments{
\item{rows}{Selected rows in the fits database. Use one of
\itemize{
\item a vector of \code{fitids}
\item 'all' for all fits
\item a named list to filter fits. List items are \verb{<field in fdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{cols}{Selected columns to display. Use one of
\itemize{
\item \emph{brief} (1)
\item \emph{normal} (2)
\item \emph{long} (3)
\item \emph{all} (4)
\item 1, 2, 3, or 4 is a shortcut for the above column sets
\item A vector of column names to include
}}

\item{report}{If TRUE, give a report (on a single fit); otherwise, list info on fits.
If rows is a numeric scalar, report defaults to TRUE; otherwise FALSE.}

\item{sort}{The name of the column to be used to sort the table}

\item{decreasing}{If TRUE, sort in descending order}

\item{nrows}{Number of rows to display in the table. Positive numbers
display the first \emph{n} rows, and negative numbers display the last \emph{n} rows.
Use \code{nrows = NA} to display all rows.}

\item{quiet}{If TRUE, doesn't print anything, just returns values}

\item{timezone}{Time zone for launch time; use NULL to leave times in native UTC}
}
\value{
The fit table or assessment, invisibly
}
\description{
Display model specification, assessment, and run statistics.
}
\details{
\code{fitinfo} works in three different modes:
\itemize{
\item \verb{fitinfo(rows = <selected rows>, cols = <selected columns>)} displays a table
of selected rows and columns
\item \verb{fitinfo(rows = <a single row>)} \emph{or} \code{fitinfo(rows = ..., report = TRUE)} displays
a report for the selected fit id, focusing on the model assessment (the same
information in the \code{fit} log)
}
}
