% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpurge.R
\name{fitpurge}
\alias{fitpurge}
\title{Purge selected fits from fits database}
\usage{
fitpurge(rows = NULL, failed = FALSE, files = FALSE, undo = FALSE)
}
\arguments{
\item{rows}{Selected rows in the fits database. Use one of
\itemize{
\item a vector of \code{fitids}
\item a named list to filter fits. List items are \verb{<field in fdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{failed}{If TRUE, all fits where \code{success} is \code{FALSE} or \code{NA} are purged}

\item{files}{If TRUE, deletes all files associated with fits that are not in database}

\item{undo}{If TRUE, the previous purge call is reversed, and the database and
associated files are restored. Call \code{fitpurge(undo = TRUE)} repeatedly to roll
back farther.}
}
\description{
Purged rows in the fits database are saved to \code{databases/purged/fdb_purged.RDS}, and
associated files are moved to \verb{database/purged/}.
}
