% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitset.R
\name{fitset}
\alias{fitset}
\title{Set model scores and comments}
\usage{
fitset(rows, score = NULL, assess = NULL, map = NULL, multiple = FALSE)
}
\arguments{
\item{rows}{Selected rows in the fits database. Use one of
\itemize{
\item a vector of \code{fitids}
\item 'all' for all fits
\item a named list to filter fits. List items are \verb{<field in fdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{score}{Set the subjective model score in the fits database. This may be numeric
or character; it'll be treated as a character.}

\item{assess}{Sets the assessment comment in the fits database}

\item{map}{Sets the map comment in the fits database}

\item{multiple}{If TRUE, allows applying to multiple fits}
}
\description{
\itemize{
\item Any of

\if{html}{\out{<div class="sourceCode">}}\preformatted{fitset(rows = <a single row>, score = <fit score>)
fitset(rows = <a single row>, assess = 'an assessment comment')
fitset(rows = <a single row>, map = 'a map comment')
}\if{html}{\out{</div>}}

sets the subjective model fit score, the assessment. comment, or the map
comment. Any of these may be combined in a single call. Note that you can
use \code{fitset} on multiple fits, but you'll need to use \code{multiple = TRUE}.
}
}
