% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_POSIXct.R
\name{fix_POSIXct}
\alias{fix_POSIXct}
\title{Repair mixed POSIXct/numeric vectors}
\usage{
fix_POSIXct(x)
}
\arguments{
\item{x}{Vector that should all be POSIXct}
}
\value{
Vector that all is POSIXct
}
\description{
I don't know how these happen, but they keep showing up. Strings look like this

\if{html}{\out{<div class="sourceCode">}}\preformatted{"2025-10-04 01:22:44.55288"  "2025-10-04 01:22:44.55288"  "1767713663.71366"  "1767713663.71366" 
}\if{html}{\out{</div>}}
}
\keyword{internal}
