% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightinfo.R
\name{flightinfo}
\alias{flightinfo}
\title{Summarize orthos at a single site}
\usage{
flightinfo(site, derived = FALSE)
}
\arguments{
\item{site}{Three letter site code}

\item{derived}{If TRUE, include derived variables}
}
\description{
For each year, for each type + sensor, for each season, for each tide stage,
lists names of orthos (\%missing).
}
