% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_image.R
\name{flights_image}
\alias{flights_image}
\title{Create cached images of orthophotos for display in \code{screen}}
\usage{
flights_image(data, cdir, file, rgb, extent, footprint, pixels = 1200)
}
\arguments{
\item{data}{raster of ortho image}

\item{cdir}{cache directory}

\item{file}{File name to write cache to}

\item{rgb}{RGB bands, in order (reversed for Mica)}

\item{extent}{Extent of image; one of \code{full}, \code{inset1}, or \code{inset2}}

\item{footprint}{footprint shapefile object}

\item{pixels}{Maximum resolution in pixels}
}
\description{
Create cached images of orthophotos for display in \code{screen}
}
\keyword{internal}
