% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_prep.R
\name{flights_prep}
\alias{flights_prep}
\title{Prepare flights data after \code{gather}}
\usage{
flights_prep(site)
}
\arguments{
\item{site}{site, using 3 letter abbreviation}
}
\description{
Run by \code{gather}, this
\enumerate{
\item Gets percent missing for each ortho (checks for updated orthos from
missing_filestamp in flights database)
\item Makes a raster of number of orthos with a missing value in each cell
(redoes this if any orthos change)
\item Gets number of bands for each ortho
\item Caches images for \code{screen} (checks to see if images are outdated with
respect to orthos)
The flights database is updated accordingly.
}
}
\keyword{internal}
