% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_report.R
\name{flights_report}
\alias{flights_report}
\title{Produce a flights report for one or more sites}
\usage{
flights_report(site = "all")
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. Use \code{all} to process all sites.}
}
\description{
The site report is a tab-delimited text file listing images flagged for repair.

Files are written to \code{reports/flagged_for_repair.txt}.
}
