% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_finish.R
\name{gather_finish}
\alias{gather_finish}
\title{Finish gather run}
\usage{
gather_finish(jobid, status)
}
\arguments{
\item{jobid}{Job ids to finish for}

\item{status}{Job status}
}
\description{
Copies the log file to the flights directory
}
\keyword{internal}
