% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flights_db.R
\name{get_flights_db}
\alias{get_flights_db}
\title{Get the flights database for a site}
\usage{
get_flights_db(site, noerror = FALSE)
}
\arguments{
\item{site}{A single site name (3 letter code), already vetted}

\item{noerror}{If TRUE, return NULL instead of error}
}
\value{
the flights database, with deleted files excluded
}
\description{
Get the flights database for a site
}
\keyword{internal}
