% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NAflag.R
\name{get_NAflag}
\alias{get_NAflag}
\title{Returns the \code{NAflag} for a \code{SpatRaster}, making one up if it's not defined}
\usage{
get_NAflag(rast)
}
\arguments{
\item{rast}{A \code{SpatRaster} object}
}
\value{
The existing or assigned \code{NAflag}
}
\description{
Returns the \code{NAflag} for a \code{SpatRaster}, making one up if it's not defined
}
\keyword{internal}
