% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rast.R
\name{get_rast}
\alias{get_rast}
\title{A wrapper for rast(get_file) that sets missing values to NA}
\usage{
get_rast(name, gd, logfile)
}
\arguments{
\item{name}{File path and name}

\item{gd}{Source drive info, named list of
\itemize{
\item \code{dir} - Google directory info, from \link{get_dir}
\item \code{sourcedrive} - which source drive (\code{local}, \code{google}, or \code{sftp})
\item \code{sftp} - list(url, user)
\item \code{cachedir} - local cache directory
}}

\item{logfile}{Log file, for reporting missing directories (which don't throw an error)}
}
\value{
path to file on local drive
}
\description{
The image geoTIFFs for the Salt Marsh project don't have NAflag set, leading
to trouble downstream. This function reads a raster from the Google Drive,
SFTP, or local drive, and if NAflag isn't set, comes up with an NAflag based
on the data type of the raster, (most commonly in our use, 255 for unsigned
bytes and 65535 for unsigned 32-bit integers) and sets these values to NA.
}
\details{
See \link{get_file} for more info.
}
\keyword{internal}
