% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Get names of one or more sites}
\usage{
get_sites(site)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. Use \code{all} to process all sites.}
}
\value{
site Vector of one or more standard 3 letter site abbreviations
}
\description{
Get names of one or more sites
}
\keyword{internal}
