% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Produce a summary of variable importance across multiple fits}
\usage{
importance(ids = NULL)
}
\arguments{
\item{ids}{Vector of fit ids, or NULL to run for all finished fits}

\item{constrain}{A list of attributes to constrain images, e.g.,
\code{contrain = list(season = 'summer', tide = c('low', 'mid'))}. This would
only include images captured in summer with low or mid-tide.}
}
\description{
Produce a summary of variable importance across multiple fits
}
