% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketiles.R
\name{maketiles}
\alias{maketiles}
\title{Create tiles with numbered blocks}
\usage{
maketiles(site, sizes)
}
\arguments{
\item{site}{Site to align tiles with}

\item{sizes}{Tile size (m); may be a vector to create several sets of tiles}
}
\description{
Create one or more shapefiles corresponding to the named site with tiles of the
specified sizes, numbered in square blocks of 1-9. Used with the \code{blocks} option
in \code{fit} to reduce the effects of spatial autocorrelation.
}
