% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_finish.R
\name{map_finish}
\alias{map_finish}
\title{Finish do_map run, updating maps database}
\usage{
map_finish(jobid, status)
}
\arguments{
\item{jobid}{slurmcollie job id}

\item{status}{Job status}
}
\description{
Finish do_map run, updating maps database
}
