% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapinfo.R
\name{mapinfo}
\alias{mapinfo}
\title{Display information on maps}
\usage{
mapinfo(
  rows = "all",
  cols = "normal",
  sort = "mapid",
  decreasing = FALSE,
  nrows = NA,
  quiet = FALSE,
  timezone = "America/New_York"
)
}
\arguments{
\item{rows}{Selected rows in the maps database. Use one of
\itemize{
\item a vector of \code{mapids}
\item 'all' for all maps
\item a named list to filter maps. List items are \verb{<field in mdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{cols}{Selected columns to display. Use one of
\itemize{
\item \emph{brief} (1)
\item \emph{normal} (2)
\item \emph{long} (3)
\item \emph{all} (4)
\item 1, 2, 3, or 4 is a shortcut for the above column sets
\item A vector of column names to include
}}

\item{sort}{The name of the column to be used to sort the table}

\item{decreasing}{If TRUE, sort in descending order}

\item{nrows}{Number of rows to display in the table. Positive numbers
display the first \emph{n} rows, and negative numbers display the last \emph{n} rows.
Use \code{nrows = NA} to display all rows.}

\item{quiet}{If TRUE, doesn't print anything, just returns values}

\item{timezone}{Time zone for launch time; use NULL to leave times in native UTC}
}
\value{
The model table, invisibly
}
\description{
Display map specification and run statistics.
}
