% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappurge.R
\name{mappurge}
\alias{mappurge}
\title{Purge selected maps from maps database}
\usage{
mappurge(rows = NULL, failed = FALSE, undo = NULL)
}
\arguments{
\item{rows}{Selected rows in the maps database. Use one of
\itemize{
\item an empty string doesn't purge any maps, but does purge stray map files
\item a vector of \code{mapids} to purge those maps
\item a named list to filter maps. List items are \verb{<field in mdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{failed}{If TRUE, all maps where \code{success} is \code{FALSE} or \code{NA} are purged.
This is an alternative to specifying \code{rows}.}

\item{undo}{Undo previous purges. There is no time limit on undoing, and it doesn't
matter whether you've run more maps since a purge. You may supply either:
\itemize{
\item \code{undo = 'last'}, reverses the previous purge call, and the
database and associated files are restored. Call \code{mappurge(undo = 'last')}
repeatedly to roll back farther.
\item a vector of map ids corresponding to previously purged maps.
Note that you may view purged maps with \verb{mapinfo(..., purged = TRUE}).
}}
}
\description{
Cleans up the maps database of unwanted maps, \emph{\strong{as well as cleaning up sidecar
files}}. Purges are always reversible.
}
\details{
Purged rows in the maps database are saved to \code{databases/purged/mdb_purged.RDS},
and associated files are moved to \verb{databases/purged/}.

You can undo previous purges or unpurge specific maps with \code{undo}. Use
\code{mapinfo(..., purged = TRUE)} to view purged maps.

If you're hurting for disk space, it's fair game to delete
\code{databases/purged/mdb_purged.RDS} and files in \verb{databases/purged/}. Doing so
will prevent you from recovering purged maps, of course.

\emph{\strong{Do I want purging maps to delete TIFFs?!}}
}
