% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_db.R
\name{new_db}
\alias{new_db}
\title{Create a new database}
\usage{
new_db(database, really = FALSE)
}
\arguments{
\item{database}{Name of database (\code{fdb} or \code{mdb})}

\item{really}{If TRUE, creates database, \strong{destroying existing database}}
}
\description{
Creates a new empty fits database (\code{fdb}) or maps database (\code{mdb}). This is a
drastic function, intended to be used only when initially creating a database
or when an existing database is a hopeless mess. Use with great care--this
function will destroy any existing database and backups. \\strong{This function is drastic and
unrecoverable.}
}
