% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{overlaps}
\alias{overlaps}
\title{Erase overlapping polys where class field isn't equal}
\usage{
overlaps(polys, field, all = TRUE)
}
\arguments{
\item{polys}{sf object with potential multipolys}

\item{field}{Field to check for overlaps}

\item{all}{If TRUE, erase all overlaps, even if fields are equal}
}
\value{
New sf object with overlaps erased unless all values in field are equal
}
\description{
Erase overlapping polys where class field isn't equal
}
\keyword{internal}
