% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{parse_image_filename}
\alias{parse_image_filename}
\title{Parse an image filename into components}
\usage{
parse_image_filename(filename)
}
\arguments{
\item{filename}{Image filename like "01Sep23_NOR_Mid_Mica_Ortho.tif" or
"01Sep23_NOR_Mid_Mica_Ortho__NDVI.tif" or "01Sep23_NOR_Mid_Mica_DEM.tif"}
}
\value{
List with date, site, tide, sensor, layer_type
}
\description{
Parse an image filename into components
}
\keyword{internal}
