% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Pick the best image from a portable name}
\usage{
pick(portable, db)
}
\arguments{
\item{portable}{Portable name to find}

\item{db}{Flights database}
}
\value{
Row in the database with the chosen image
}
\description{
Given a portable name and flights database, pick the image matching the
portable name with the highest score. Ties are broken by picking the earliest
day in the season. In the unlikely event there is more than one high-scoring
flight in the the first one in the database is arbitrarily chosen.
}
\keyword{internal}
