% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_bytes.R
\name{pixel_bytes}
\alias{pixel_bytes}
\title{Get the number of bytes in a pixel of a SpatRaster}
\usage{
pixel_bytes(rast)
}
\arguments{
\item{rast}{A SpatRaster}
}
\value{
The number of bytes in a pixel
}
\description{
Get the number of bytes in a pixel of a SpatRaster
}
