% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot_predict}
\alias{plot_predict}
\title{Read a predict grid and plot it nicely}
\usage{
plot_predict(
  tiff = "C:/Work/saltmarsh/data/oth/predicted/predict_oth_2025-Apr-28_13-54.tif",
  pal = "ggsci::category20b_d3"
)
}
\arguments{
\item{tiff}{A predicted geoTIFF}

\item{pal}{Selected palette}
}
\description{
Read a predict grid and plot it nicely
}
\keyword{internal}
