% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_unet.R
\name{prep_unet}
\alias{prep_unet}
\title{Pre-process data for U-Net}
\usage{
prep_unet(
  model,
  save_gis = FALSE,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{model}{The model name, which is also the name of a \code{.yml} parameter file in the \code{pars}
directory. This file must contain the following:
\itemize{
\item years: the year(s) of field data to fit
\item orthos: file names of all orthophotos to include
\item patch: size in pixels
\item depth: number of of downsampling stages
\item classes: vector of target classes
\item holdout: holdout set to use (uses bypoly\if{html}{\out{<holdout>}}, classes 1 and 6). Holdout sets are
created by \code{gather} to yield at least 20\% of separate polys. There are 5 sets to choose from.
\item overlap: Proportion overlap of patches
\item upscale: number of cells to upscale (default = 1). Use 3 to upscale to 3x3, 5 for 5x5, etc.
\item smooth: number of cells to include in moving window mean (default = 1). Use 3 to smooth to 3x3, etc.
}}

\item{save_gis}{If TRUE, saves GIS data for assessment and debugging}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}. These take priority
over the function's defaults.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error handling. Use this
for debugging. If you get unrecovered errors, the job won't be added to the jobs database. Has
no effect if local = FALSE.}

\item{comment}{Optional slurmcollie comment}
}
\description{
Creates numpy arrays ready for fitting in U-Net. Result files are placed in \verb{<site>/unet/<model>}.
}
\details{
Try this to test:
\code{prep_unet('unet01', local = TRUE)}
}
