% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_unet_finish.R
\name{prep_unet_finish}
\alias{prep_unet_finish}
\title{Finish prep_unet run}
\usage{
prep_unet_finish(jobid, status)
}
\arguments{
\item{jobid}{Job id to finish for}

\item{status}{Job status}
}
\description{
Finish a U-Net data preparation run:
\itemize{
\item Copy the log file to the unet model directory as \verb{prep_<model>.log}
}
}
