% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_transects_for_patch.R
\name{rasterize_transects_for_patch}
\alias{rasterize_transects_for_patch}
\title{Helper to rasterize transects for one dataset (train/val/test)}
\usage{
rasterize_transects_for_patch(transects, patch_ext, template, class_mapping)
}
\arguments{
\item{transects}{sf object of transects for this dataset}

\item{patch_ext}{terra extent for current patch}

\item{template}{terra raster template}

\item{class_mapping}{Named vector for class remapping}
}
\value{
List with mask_array, label_array, n_pixels, classes_string (or NULL if no data)
}
\description{
Helper to rasterize transects for one dataset (train/val/test)
}
\keyword{internal}
