% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_dir.R
\name{resolve_dir}
\alias{resolve_dir}
\title{Resolve directory with embedded \verb{<site>}, \verb{<SITE>}, \verb{<site_name>}, or \verb{<share>}}
\usage{
resolve_dir(dir, site, share = site_name)
}
\arguments{
\item{dir}{Directory path}

\item{site}{Site name. For Google Drive, use \code{site_name}; on Unity, use
\code{tolower(site)}, 3 letter code}

\item{share}{Share site name. For Google Drive, use \code{share}}
}
\value{
Directory path including specified site.
}
\description{
Resolve directory with embedded \verb{<site>}, \verb{<SITE>}, \verb{<site_name>}, or \verb{<share>}
}
\keyword{internal}
