% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_finish.R
\name{sample_finish}
\alias{sample_finish}
\title{Finish sample run}
\usage{
sample_finish(jobid, status)
}
\arguments{
\item{jobid}{Job ids to finish for}

\item{status}{Job status}
}
\description{
Copies the log file to the samples directory
}
\keyword{internal}
