% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_freq.R
\name{sample_freq}
\alias{sample_freq}
\title{Create a frequency table of number of cells by subclass from field data}
\usage{
sample_freq(sites, transects = NULL)
}
\arguments{
\item{sites}{One or more site names, using 3 letter abbreviation. Use \code{all} to process all sites.}

\item{transects}{Name of transects file; default is \code{transects}.}
}
\description{
Create a frequency table of number of cells by subclass in field data geoTIFF for one or more
sites. Results are written to the reports directory as \verb{sample_freq_<site>.txt}.
}
\details{
Always runs locally. Takes about a minute per site.
}
