% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleinfo.R
\name{sampleinfo}
\alias{sampleinfo}
\title{Give summary of sample data file}
\usage{
sampleinfo(site, datafile = "data")
}
\arguments{
\item{site}{Three letter site code}

\item{datafile}{Name of \code{.RDS} data file, excluding the extension}
}
\value{
Returns the full list of portable names, invisibly
}
\description{
Give summary of sample data file
}
