% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_database.R
\name{save_database}
\alias{save_database}
\title{Save the specified database}
\usage{
save_database(database)
}
\arguments{
\item{database}{Name of database (should be \code{fdb} for the model fit database, or
\code{mdb} for the map database)}
}
\description{
Saves the database in directory \code{the$dbdir}. Previous versions are renamed e.g., \code{fdb_1.RDS},
\code{fdb_2.RDS}, etc. Will need a mechanism to delete databases more than a week old (or something).
}
\details{
For the model fit database, \code{fdb}, \code{the$last_fit_id} is written to \code{last_fit_id.txt} to track
the highest fit id used, as these ids are \emph{never} reused.
}
\keyword{internal}
