% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_flights_db.R
\name{save_flights_db}
\alias{save_flights_db}
\title{Saves the flights database}
\usage{
save_flights_db(db, db_name)
}
\arguments{
\item{db}{Database table}

\item{db_name}{Path and name of database file}
}
\description{
Saves the flights database
}
\keyword{internal}
