% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_screen_db.R
\name{save_screen_db}
\alias{save_screen_db}
\title{Saves the screening database}
\usage{
save_screen_db(db, db_name)
}
\arguments{
\item{db}{Database table}

\item{db_name}{Path and name of database file}
}
\description{
Saves the screening database
}
\keyword{internal}
