% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_filter.R
\name{screen_filter}
\alias{screen_filter}
\title{Filter/refilter images}
\usage{
screen_filter(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Filter/refilter images
}
\keyword{internal}
