% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_image.R
\name{screen_image}
\alias{screen_image}
\title{Display the current image plus info}
\usage{
screen_image(score_choices, input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Display the current image plus info
}
\keyword{internal}
