% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_insets.R
\name{screen_insets}
\alias{screen_insets}
\title{Show inset images}
\usage{
screen_insets(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Show inset images
}
\keyword{internal}
