% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_no_site.R
\name{screen_no_site}
\alias{screen_no_site}
\title{When there's no info for a site, disable controls and blank out info}
\usage{
screen_no_site(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
When there's no info for a site, disable controls and blank out info
}
\keyword{internal}
