% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_site_info.R
\name{screen_site_info}
\alias{screen_site_info}
\title{Put together site info message}
\usage{
screen_site_info(sites, input, output, session)
}
\arguments{
\item{sites}{Sites table}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Put together site info message
}
\keyword{internal}
