% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_google_drive.R
\name{set_up_google_drive}
\alias{set_up_google_drive}
\title{Creates a token to set up the Google Drive}
\usage{
set_up_google_drive()
}
\description{
This needs to be run just once for each user of \code{saltmarsh}. It sends you
to a Google authorization page where you'll need to log in to your Google
account, copy a token, and paste it when you're asked to "Enter authorization code."
}
\details{
Once this is done, you can verify access with

\if{html}{\out{<div class="sourceCode">}}\preformatted{googledrive::drive_find(n_max = 5)
}\if{html}{\out{</div>}}

If you write code that needs to access the Google Drive, you'll need to call the
following once pre R session. \code{saltmarsh} takes care of this--it's only necessary
for separate code.

\if{html}{\out{<div class="sourceCode">}}\preformatted{googledrive::drive_auth(token = readRDS('~/.google_auth/google_drive_token.RDS'))
}\if{html}{\out{</div>}}
}
