% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_holdout.R
\name{spatial_holdout}
\alias{spatial_holdout}
\title{Number polygons spatially for each subclass to give spatially-reasonable holdout sets}
\usage{
spatial_holdout(shape, field = "subclass", count = 10)
}
\arguments{
\item{shape}{\code{sf} object}

\item{field}{Name of field to stratify on; use NULL for no stratification}

\item{count}{How many groups? You probably want 10.}
}
\value{
The modified \code{sf} object
}
\description{
Polygons are numbered from \code{1:count}, starting at the northwest-most centroid and taking
the closest centroid in turn. Numbering is stratified by \code{field}. Results are placed in
field \code{bypoly00} in the \code{sf} object.
}
