% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{the}
\alias{the}
\title{Environment for package parameters}
\description{
The environment \code{the} is used for user parameters. They
are assigned by \code{init()}, which is run when the package is
loaded, and also may be run by the user, e.g., when parameter
files are changed.
}
