% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train and assess U-Net model}
\usage{
train(
  model,
  train = NULL,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{model}{The base name of a \code{.yml} file in \verb{<pars>/unet/} with model parameters. This
file contains parameters used in data prep as well as training. Note that \code{prep_unet}
must be run after changing any of these parameters. The \code{model} file must contain the
following:
\itemize{
\item site: the three-letter site code
\item years: the year(s) of field data to fit
\item orthos: file names of all orthophotos to include
\item patch: size in pixels
\item depth: number of of downsampling stages
\item classes: vector of target classes (in original classification)
\item holdout_col: holdout set to use (uses bypoly\if{html}{\out{<holdout>}}). Holdout sets are created by
\code{gather}, numbering each poly from 1 to 10, repeating if necessary. There are 5 sets to
choose from.
\item cv: number of cross-validations. Use 1 for a single model, up to 5 for five-fold
cross-validation. Cross-validations are systematic, not random. Since there are only 10
sets in each bypoly, the number of cross-validations is limited by the values of val
and test.
\item val: validation polys from \code{holdout_col}. Use NULL to skip validation, or a vector of
the validation polys for the first cross-validation (these will be incremented for
subsequent validations). For 20\% validation holdout, use \code{val = c(1, 6)}. This will use
\verb{bypoly01 \%in\% c(1,6)`` for the first cross-validation, }c(2, 7)` for the second, and so
on.
\item test: test polys from \code{holdout_col}, as with \code{val}.
\item overlap: Proportion overlap of patches
\item upscale: number of cells to upscale (default = 1). Use 3 to upscale to 3x3, 5 for 5x5, etc.
\item smooth: number of cells to include in moving window mean (default = 1). Use 3 to smooth to 3x3, etc.
}}

\item{train}{The base name of a \code{.yml} file in \verb{<pars>/unet/} with training parameters.
If present, this overrides parameters in \code{model}. This file contains parameters used only
in the training phase. The following must be present either in the \code{model} or \code{train} file:
\itemize{
\item in_channels Number of input channels (8 for multispectral + NDVI + NDRE + DEM)
\item n_epochs Number of training epochs
\item encoder_name. Pre-trained encoder to use. Choices include \code{resnet10}, \code{resnet18},
\code{resnet34}, \code{resnet50}, \code{efficientnet-b0}, and others. The lower \code{restnet}
numbers have fewer parameters, so may be likely to result in more stable training.
\item encoder_weights. \code{imagenet} start with weights learned on ImageNet (natural images);
gives faster convergence, but might bias toward RGB patterns. NULL starts with random
initialization, thus learns everything from this dataset; no bias, but slower training.
\item learning_rate Learning rate for optimizer
\item weight_decay. L2 regularization - penalizes large weights to prevent overfitting.
Higher values (1e-3) = stronger regularization. Lower values (1e-5) = weaker.
\item batch_size. How many patches to process together. Larger (16, 32) uses parallelization
on GPUs so trains faster, more stable gradients, uses more GPU memory. Smaller (4, 8) gives
noisier gradients (good regularization), less memory, better for small datasets. Use 8; if
overfitting is a problem, try 4.
\item gradient_clip_max_norm. Prevents exploding gradients by capping gradient magnitude.
Range: 0.5 (aggressive clipping) to 5.0 (gentle); start with 1.0.
\item use_ordinal If TRUE, use ordinal regression U-Net
}}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}. These take priority
over the function's defaults. \strong{Note that this function requires GPUs}. By default, it
requests 2 L40S GPUs.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error handling. Use this
for debugging. If you get unrecovered errors, the job won't be added to the jobs database. Has
no effect if local = FALSE.}

\item{comment}{Optional slurmcollie comment}
}
\description{
Train a U-Net model. Result files are placed in \verb{<site>/unet/<model>}.
}
