% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_finish.R
\name{train_finish}
\alias{train_finish}
\title{Finish train run}
\usage{
train_finish(jobid, status)
}
\arguments{
\item{jobid}{Job id to finish for}

\item{status}{Job status}
}
\description{
Finish a U-Net train run:
\itemize{
\item Populate fits database with \code{slurmcollie} stats
\item and with info from \verb{zz_<id>_train.RDS}, written by \code{do_train}
\item Copy the log file to the models directory as \verb{fit_<id>.log}
\item Copy summary to \verb{fit_<id>_summary.txt}
\item Copy training curves to \verb{fit_<id>_training_curves.png} (if present)
}
}
