% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_predict.R
\name{unet_confusion_matrix}
\alias{unet_confusion_matrix}
\title{Create confusion matrix from U-Net predictions}
\usage{
unet_confusion_matrix(pred_results)
}
\arguments{
\item{pred_results}{Results from unet_predict()}
}
\value{
confusionMatrix object from caret
}
\description{
Create confusion matrix from U-Net predictions
}
