% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_export_to_numpy.R
\name{unet_export_to_numpy}
\alias{unet_export_to_numpy}
\title{Export prepared data to numpy arrays for Python}
\usage{
unet_export_to_numpy(patches, output_dir, site, class_mapping)
}
\arguments{
\item{patches}{List from unet_extract_training_patches}

\item{output_dir}{Directory to save numpy files}

\item{site}{Name for files (e.g., 'rr')}

\item{class_mapping}{Named vector mapping original to remapped classes (e.g., c('3'=0, '4'=1, '5'=2, '6'=3))}
}
\description{
Export prepared data to numpy arrays for Python
}
