% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_patch_stats.R
\name{unet_patch_stats}
\alias{unet_patch_stats}
\title{Summary stats for extracted patches}
\usage{
unet_patch_stats(patch_data)
}
\arguments{
\item{patch_data}{Extracted patches from \code{unet_extract_training_patches}}
}
\value{
Data frame of stats for each patch (\code{patch_id}, \code{n_labeled}, \code{n_classes},
\code{dominant_class}, \code{purity})
}
\description{
Hope to see single class patches < 80\%.
}
\keyword{internal}
