% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_prep_finish.R
\name{unet_prep_finish}
\alias{unet_prep_finish}
\title{Finish unet_prep run}
\usage{
unet_prep_finish(jobid, status)
}
\arguments{
\item{jobid}{Job id to finish for}

\item{status}{Job status}
}
\description{
Finish a U-Net data preparation run:
\itemize{
\item Copy the log file to the unet model directory as \verb{prep_<model>.log}
}
}
