% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_spatial_train_val_split.R
\name{unet_spatial_train_val_split}
\alias{unet_spatial_train_val_split}
\title{Split transects into train and validation sets}
\usage{
unet_spatial_train_val_split(transects, holdout)
}
\arguments{
\item{transects}{Original sf transects object}

\item{holdout}{Holdout set to use (uses bypoly\if{html}{\out{<holdout>}}, classes 1 and 6)}
}
\value{
List with train and val transect ids
}
\description{
Split transects into train and validation sets
}
\keyword{internal}
