% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_clone.R
\name{upscale_clone}
\alias{upscale_clone}
\title{Clones a site directory at a new grain}
\usage{
upscale_clone(site, newsite, cellsize)
}
\arguments{
\item{site}{Site name}

\item{newsite}{Name for cloned site}

\item{cellsize}{Cell size for new site (m)}
}
\description{
\itemize{
\item this is experimental, to see if true rescaling helps
\item you'll have to add site to sites.txt on your own \strong{before} running this
\item there is currently no path to use \code{gather} on the cloned site;
I'll add it if this seems promising
\item upscaled orthos are not cloned
\item rejected orthos are not cloned
\item scores and comments are copied to the new site
\item this doesn't bother with blocks files
\item this only affects the \code{flights} directory. You should be able
to run \code{screen}, \code{derive}, \code{sample}, and other functions
normally after cloning.
\item I could add \code{sd} as an option, making multiple results. You can
also specify arbitrary functions, so this would be a good place
for other upscaling funs.
}
}
