% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_more.R
\name{upscale_more}
\alias{upscale_more}
\title{Adds more metrics to an already existing upscaled clone}
\usage{
upscale_more(site, newsite, cellsize, metrics = "all")
}
\arguments{
\item{site}{Site name}

\item{newsite}{Name for cloned site}

\item{cellsize}{Cell size for new site (m)}

\item{metrics}{A list of metrics, or 'all' for all metrics. May include any of:
\itemize{
\item \code{sd} Standard deviation
\item \code{q05}, \code{q10}, \code{q25}, \code{median}, \code{q75}, \code{q90}, and \code{q95} Quantiles
\item \code{r0595}, \code{r1090}, \code{iqr} Quantile ranges: 5th-95th, 10th-90th, and interquartile range
\item \code{skewness} and \code{kurtosis}, for Ryan
}}
}
\description{
\itemize{
\item this is experimental
}
}
