% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_layer.R
\name{write_layer_sidecar}
\alias{write_layer_sidecar}
\title{Write the sidecar YAML documenting the integration}
\usage{
write_layer_sidecar(
  sidecar_path,
  cfg,
  final_r,
  source_id_rast,
  source_maps,
  class_lookup
)
}
\description{
Write the sidecar YAML documenting the integration
}
\keyword{internal}
