% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{Create derived variables such as NDVI and NDRE}
\usage{
derive(
  site,
  pattern1 = "",
  pattern2 = NULL,
  metrics = c("NDVI", "NDWIg", "NDRE"),
  window = 3,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. Use \code{all} to process all sites.
In batch mode, each named site will be run in a separate job.}

\item{pattern1}{File names, portable names, regex matching either, or search names
selecting source for derived variables. See Image naming in
\href{https://github.com/UMass-UAS-Salt-Marsh/salt-marsh-mapping/blob/main/README.md}{README}
for details. See details.}

\item{pattern2}{A second pattern or vector of layer names, used for bivariate metrics. See details.}

\item{metrics}{A list of metrics to apply. Univariate metrics include:
\describe{
\item{NDVI}{Normalized difference vegetation index, \code{(NIR - red) / (NIR + red)}, an index of biomass}
\item{NDWIg}{Normalized difference water index (green, commonly known as McFeeter's \code{NDWI}),
\code{(green - NIR) / (green + NIR)}, primarily helps distinguish waterbodies}
\item{NDRE}{Normalized difference red edge index, \code{(NIR - RE) / (NIR + RE)}, an index of the
amount of chlorophyll in a plant}
\item{mean}{mean of each band in a window, size defined by \code{window}}
\item{std}{standard deviation of each band in a window, size defined by \code{window}}
\item{NDVImean}{mean of NDVI in a window, size defined by \code{window}}
\item{NDVIstd}{standard deviation of NDVI in a window, size defined by \code{window}}
Bivariate metrics include:
\item{NDWIswir}{Normalized difference water index (SWIR, commonly known as Gao's \code{NDWI}),
\code{(NIR - SWIR) / (NIR + SWIR)}, an index of water content in leaves; requires a Mica layer for
\code{pattern1}, and a matched SWIR layer for \code{pattern2}}
\item{delta}{The difference between \code{pattern1} and \code{pattern2}, may be useful for taking a
difference between late-season and early-season DEMs to represent vegetation canopy height}
}}

\item{window}{Window size for \code{mean}, \code{std}, \code{NDVImean}, and \code{NDVIstd}, in cells; windows are square, so just specify
a single number. Bonus points if you remember to make it odd.}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}. These take priority
#'    over the function's defaults.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error handling. Use this
for debugging. If you get unrecovered errors, the job won't be added to the jobs database. Has
no effect if local = FALSE.}

\item{comment}{Optional slurmcollie comment}
}
\description{
This function creates any of a number of univariate and bivariate metrics derived from
raster data, such as NDVI and NDRE. Results are written as rasters to the \code{flights} directory,
with the metric included in the name. These metrics will be treated like source layers in
subsequent processing and modeling.
}
\details{
For univariate metrics, supply one or more layer names via \code{pattern1}. All metrics will be
calculated for each layer specified by \code{pattern1}. Results will be named \verb{<layer>__<metric>}.

For bivariate metrics, specify matched pairs of layers with \code{pattern1} and \code{pattern2}. It's
best to specify complete names (you can use vectors for each) so the layers are paired properly.
If you're crazy enough to use regular expressions here, scrutinize the result names carefully.
Results will be named \verb{<layer1>__<layer2>__metric}. At the moment, \code{NDWIswir} and \code{delta} are the
only bivariate metrics.

Note that all normalized difference (\code{NDxx}) metrics require five-band Mica data.

Note that derived metrics get two underscores in their names, e.g., \verb{<layer>__NDVI}. This is
used to distinguish primary from derived data.

This fits in the workflow after \code{gather} and before \code{sample}.
}
