% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_derive.R
\name{do_derive}
\alias{do_derive}
\title{Create derived variables such as NDVI and NDRE}
\usage{
do_derive(
  site,
  pattern1 = "mica",
  pattern2 = NULL,
  metrics = c("NDVI", "NDRE"),
  window = 3
)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. If running in batch mode, each
named site will be run in a separate job.}

\item{pattern1}{Regex filtering rasters, case-insensitive. Default = "" (match all). Note: only
files ending in \code{.tif} are included in any case. \code{pattern1} may alternatively be a vector
of layer names (or patterns if you want to get tricky).}

\item{pattern2}{A second pattern or vector of layer names, used for bivariate metrics. See details.}

\item{metrics}{A list of metrics to apply. Univariate metrics include:
\describe{
\item{NDVI}{Normalized difference vegetation index, \code{(NIR - red) / (NIR + red)}, an index of biomass}
\item{NDWIg}{Normalized difference water index (green, commonly known as McFeeter's \code{NDWI}), \code{(green - NIR) / (green + NIR)},
primarily helps distinguish waterbodies}
\item{NDRE}{Normalized difference red edge index, \code{(NIR - RE) / (NIR + RE)}, an index of the
amount of chlorophyll in a plant}
\item{mean}{mean of each band in a window, size defined by \code{window}}
\item{std}{standard deviation of each band in a window, size defined by \code{window}}
\item{NDVImean}{mean of NDVI in a window, size defined by \code{window}}
\item{NDVIstd}{standard deviation of NDVI in a window, size defined by \code{window}}
Bivariate metrics include:
\item{NDWIswir}{Normalized difference water index (SWIR, commonly known as Gao's \code{NDWI}), \code{(NIR - SWIR) / (NIR + SWIR)},
an index of water content in leaves; requires a Mica layer for \code{pattern1}, and a matched
SWIR layer for \code{pattern2}}
\item{delta}{The difference between \code{pattern1} and \code{pattern2}, may be useful for taking a
difference between late-season and early-season DEMs to represent vegetation canopy height}
}}

\item{window}{Window size for \code{mean}, \code{std}, \code{NDVImean}, and \code{NDVIstd}, in cells; windows are square, so just specify
a single number. Bonus points if you remember to make it odd.}
}
\description{
Create derived variables such as NDVI and NDRE
}
