% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fit.R
\name{do_fit}
\alias{do_fit}
\title{Fit models}
\usage{
do_fit(
  fitid,
  sites,
  name,
  method,
  fitargs,
  vars,
  exclude_vars,
  exclude_classes,
  min_class,
  reclass,
  max_samples,
  years,
  minscore,
  maxmissing,
  max_miss_train,
  top_importance,
  holdout,
  blocks,
  auc,
  hyper,
  rep = NULL
)
}
\arguments{
\item{fitid}{Fit id in the fits database}

\item{sites}{Data frame with \code{site} (3 letter code), \code{site_name} (long name), and
\code{datafile} (resolved path and filename of datafile). Sites, paths, and filenames
are vetted by fit - there's no checking here.}

\item{name}{Optional model name}

\item{method}{One of \code{rf} for Random Forest, \code{boost} for AdaBoost. Default = \code{rf}.}

\item{fitargs}{A named list of additional arguments to pass to the model (\code{ranger} or \code{boost})}

\item{vars}{Vector of variables to restrict analysis to. Default = \verb{\{*\}},
all variables. \code{vars} is processed by \code{find_orthos}, and may include file names,
portable names, search names and regular expressions of file and portable names.}

\item{exclude_vars}{An optional vector of variables to exclude. As with \code{vars}, variables
are processed by \code{find_orthos}}

\item{exclude_classes}{Numeric vector of subclasses to exclude}

\item{min_class}{Minimum number of training samples to allow in a class. All classes with
fewer samples in training set as well as all classes with zero cases in the
validation set will be dropped from the model. Use \code{min_class = NULL} to prevent
dropping any classes.}

\item{reclass}{Vector of paired classes to reclassify, e.g., \code{reclass = c(13, 2, 3, 4)}
would reclassify all 13s to 2 and 3s to 4, lumping each pair of classes.}

\item{max_samples}{Maximum number of samples to use - subsample if necessary}

\item{years}{Vector of years to restrict variables to}

\item{minscore}{Minimum score for orthos. Files with a minimum score of less than
this are excluded from results. Default is 0, but rejected orthos are always
excluded.}

\item{maxmissing}{Maximum percent missing in orthos. Files with percent missing greater
than this are excluded.}

\item{max_miss_train}{Maximum proportion of missing training points allowed before a
variable is dropped}

\item{top_importance}{Number of variables to keep for variable importance}

\item{holdout}{Proportion of points to hold out. For Random Forest, this specifies
the size of the single validation set, while for boosting, it is the size of each
of the testing and validation sets.}

\item{blocks}{An alternative to holding out random points. Specify a named list
with \verb{block = <name of block column>, classes = <vector of block classes to hold out>}.
Set this up by creating a shapefile corresponding to ground truth data with a variable
\code{block} that contains integer block classes, and placing it in the \verb{blocks/} directory
for the site. \code{gather} and \code{sample} will collect and process block data for you to
use here.}

\item{auc}{If TRUE, calculate class probabilities so we can calculate AUC}

\item{hyper}{Hyperparameters \emph{\strong{To be defined}}}

\item{rep}{Throwaway argument to make \code{slurmcollie} happy}
}
\description{
Fit models
}
