% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fit.R
\name{do_fit}
\alias{do_fit}
\title{Fit models}
\usage{
do_fit(
  fitid,
  sites,
  name,
  method,
  vars,
  exclude_vars,
  exclude_classes,
  years,
  minscore,
  maxmissing,
  max_miss_train,
  top_importance,
  holdout,
  auc,
  hyper,
  rep = NULL
)
}
\arguments{
\item{fitid}{Fit id in the fits database}

\item{sites}{Data frame with \code{site} (3 letter code), \code{site_name} (long name), and
\code{datafile} (resolved path and filename of datafile). Sites, paths, and filenames
are vetted by fit - there's no checking here.}

\item{name}{Optional model name}

\item{method}{One of \code{rf} for Random Forest, \code{boost} for AdaBoost. Default = \code{rf}.}

\item{vars}{An optional vector of variables to restrict analysis to. Default = \verb{\{*\}},
all variables. \code{vars} is processed by \code{find_orthos}, and may include file names,
portable names, search names and regular expressions of file and portable names.}

\item{exclude_vars}{An optional vector of variables to exclude. As with \code{vars}, variables
are processed by \code{find_orthos}}

\item{exclude_classes}{An optionial numeric vector of subclasses to exclude}

\item{years}{An optional vector of years to restrict variables to}

\item{minscore}{Minimum score for orthos. Files with a minimum score of less than
this are excluded from results. Default is 0, but rejected orthos are always
excluded.}

\item{maxmissing}{Maximum percent missing in orthos. Files with percent missing greater
than this are excluded.}

\item{max_miss_train}{Maximum proportion of missing training points allowed before a
variable is dropped}

\item{top_importance}{Number of variables to keep for variable importance}

\item{holdout}{Proportion of points to hold out. For Random Forest, this specifies
the size of the single validation set, while for boosting, it is the size of each
of the testing and validation sets.}

\item{auc}{If TRUE, calculate class probabilities so we can calculate AUC}

\item{hyper}{Hyperparameters \emph{\strong{To be defined}}}

\item{rep}{Throwaway argument to make \code{slurmcollie} happy}
}
\description{
Fit models
}
