% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_prep_unet.R
\name{do_prep_unet}
\alias{do_prep_unet}
\title{Pre-process data for U-Net}
\usage{
do_prep_unet(model)
}
\arguments{
\item{model}{The model name, which is also the name of a \code{.yml} parameter file in the \code{pars}
directory. This file must contain the following:
\itemize{
\item year: the year to fit
\item orthos: portable names of all orthophotos to include
\item patch: size in pixels
\item depth: number of of downsampling stages
\item classes: vector of target classes
\item holdout: percent of data to hold out for validation
}}
}
\description{
Creates numpy arrays ready for fitting in U-Net. Result files are placed in \verb{<site>/unet/<model>}.
}
