% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_area.R
\name{extent_area}
\alias{extent_area}
\title{Give the area of a clip vector}
\usage{
extent_area(clip, units = "ha", crs = "epsg:4326")
}
\arguments{
\item{clip}{Vector of \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}}

\item{units}{Area units to return}

\item{crs}{Coordinate reference system of clip}
}
\description{
Give the area of a clip vector
}
\keyword{internal}
