% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldinfo.R
\name{fieldinfo}
\alias{fieldinfo}
\title{Summarize number of field-sampled polys and pixels by class and year}
\usage{
fieldinfo(site)
}
\arguments{
\item{site}{Three letter site code}
}
\description{
Summarize number of field-sampled polys and pixels by class and year
}
