% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Build statistical models of vegetation cover}
\usage{
fit(
  site = NULL,
  datafile = "data",
  name = "",
  method = "rf",
  vars = "{*}",
  exclude_vars = "",
  exclude_classes = NULL,
  reclass = c(13, 2),
  max_samples = NULL,
  years = NULL,
  minscore = 0,
  maxmissing = 20,
  max_miss_train = 0.2,
  top_importance = 20,
  holdout = 0.2,
  blocks = NULL,
  auc = FALSE,
  hyper = NULL,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{site}{Three letter site code, or vector of site names if fitting multiple sites}

\item{datafile}{Name of data file. It must be an \code{.RDS} file, but exclude the
extension. If fitting multiple sites, either use a single datafile name
shared among sites, or a vector matching site.}

\item{name}{Optional model name}

\item{method}{One of \code{rf} for Random Forest, \code{boost} for AdaBoost. Default = \code{rf}.}

\item{vars}{Vector of variables to restrict analysis to. Default = \verb{\{*\}},
all variables. \code{vars} is processed by \code{find_orthos}, and may include file names,
portable names, search names and regular expressions of file and portable names.}

\item{exclude_vars}{An optional vector of variables to exclude. As with \code{vars}, variables
are processed by \code{find_orthos}}

\item{exclude_classes}{Numeric vector of subclasses to exclude}

\item{reclass}{Vector of paired classes to reclassify, e.g., \code{reclass = c(13, 2, 3, 4)}
would reclassify all 13s to 2 and 4s to 3, lumping each pair of classes.}

\item{max_samples}{Maximum number of samples to use - subsample if necessary}

\item{years}{Vector of years to restrict variables to}

\item{minscore}{Minimum score for orthos. Files with a minimum score of less than
this are excluded from results. Default is 0, but rejected orthos are always
excluded.}

\item{maxmissing}{Maximum percent missing in orthos. Files with percent missing greater
than this are excluded.}

\item{max_miss_train}{Maximum proportion of missing training points allowed before a
variable is dropped}

\item{top_importance}{Number of variables to keep for variable importance}

\item{holdout}{Proportion of points to hold out. For Random Forest, this specifies
the size of the single validation set, while for boosting, it is the size of each
of the testing and validation sets.}

\item{blocks}{An alternative to holding out random points. Specify a named list
with \verb{block = <name of block column>, classes = <vector of block classes to hold out>}.
Set this up by creating a shapefile corresponding to ground truth data with a variable
\code{block} that contains integer block classes, and placing it in the \verb{blocks/} directory
for the site. \code{gather} and \code{sample} will collect and process block data for you to
use here.}

\item{auc}{If TRUE, calculate class probabilities so we can calculate AUC}

\item{hyper}{Hyperparameters. \emph{\strong{To be defined.}}}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}.
These take priority over the function's defaults.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error
handling. Use this for debugging. If you get unrecovered errors, the job
won't be added to the jobs database. Has no effect if local = FALSE.}

\item{comment}{Optional launch / slurmcollie comment}
}
\description{
Given one or more sites and a model specification, builds a model of vegetation
cover and report model assessment.
}
