% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Build statistical models of vegetation cover}
\usage{
fit(
  site = NULL,
  datafile = "data",
  name = "",
  method = "rf",
  vars = "{*}",
  exclude = "",
  years = NULL,
  maxmissing = 0.05,
  top_importance = 20,
  holdout = 0.2,
  auc = TRUE,
  hyper = NULL,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{site}{Site name, or vector of site names if fitting multiple sites.}

\item{datafile}{Name of data file. It must be an \code{.RDS} file, but exclude the
extension. If fitting multiple sites, either use a single datafile name
shared among sites, or a vector matching site.}

\item{name}{Optional model name}

\item{method}{One of \code{rf} for Random Forest, \code{boost} for AdaBoost. Default = \code{rf}.}

\item{vars}{An optional vector of variables to restrict analysis to. Default = {*},
all variables. \code{vars} is processed by \code{find_orthos}, and may include file names,
portable names, search names and regular expressions of file and portable names.}

\item{exclude}{An optional vector of variables to exclude. As with \code{vars}, variables
are processed by \code{find_orthos}.}

\item{years}{An optional vector of years to restrict variables to.}

\item{maxmissing}{Maximum proportion of missing training points allowed before a
variable is dropped.}

\item{top_importance}{Number of variables to keep for variable importance}

\item{holdout}{Proportion of points to hold out. For Random Forest, this specifies
the size of the single validation set, while for boosting, it is the size of each
of the testing and validation sets.}

\item{auc}{If TRUE, calculate class probabilities so we can calculate AUC.}

\item{hyper}{Hyperparameters. \emph{\strong{To be defined.}}}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}.
These take priority
#'    over the function's defaults.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error
handling. Use this for debugging. If you get unrecovered errors, the job
won't be added to the jobs database. Has no effect if local = FALSE.}

\item{comment}{Optional launch / slurmcollie comment}
}
\description{
Given one or more sites and a model specification, builds a model of vegetation
cover and report model assessment.
}
