% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitinfo.R
\name{fitinfo}
\alias{fitinfo}
\title{Give info on models that have been built with \code{fit}}
\usage{
fitinfo()
}
\arguments{
\item{rows}{Selected rows in the fits database}

\item{cols}{Selected columns to display}
}
\description{
Give info on models that have been built with \code{fit}
}
