% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitinfo.R
\name{fitinfo}
\alias{fitinfo}
\title{Displays information on model fits, including model specification, assessment, and run statistics}
\usage{
fitinfo()
}
\arguments{
\item{rows}{Selected rows in the fits database}

\item{cols}{Selected columns to display}
}
\description{
Displays information on model fits, including model specification, assessment, and run statistics
}
