% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightinfo.R
\name{flightinfo}
\alias{flightinfo}
\title{Summarize orthos at a single site}
\usage{
flightinfo(site, filter = NULL, derived = FALSE)
}
\arguments{
\item{site}{Three letter site code}

\item{filter}{A named list restricting to particular values, e.g.,
\code{filter = list(type = 'DEM', season = c('spring', 'summer'))}}

\item{derived}{If TRUE, include derived variables}
}
\description{
For each year, for each type + sensor, for each season, for each tide stage,
lists names of orthos (\%missing).
}
