% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_report.R
\name{flights_report}
\alias{flights_report}
\title{Produce a flights report for one or more sites}
\usage{
flights_report()
}
\description{
The site report is series of tab-delimited text files:
}
\details{
\enumerate{
\item Summary of orthos for each site
\item List of orthos flagged for repair
\item List of duplicated portable names (* marks selected names)
}

Files are written to the \verb{reports/} directory.
}
