% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NAflag.R
\name{get_NAflag}
\alias{get_NAflag}
\title{Returns the NAflag for a SpatRaster, making one up if it's not defined}
\usage{
get_NAflag(rast)
}
\arguments{
\item{rast}{A SpatRaster object}
}
\value{
The existing or assigned NAflag
}
\description{
Returns the NAflag for a SpatRaster, making one up if it's not defined
}
\keyword{internal}
