% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Produce a summary of variable importance across multiple fits}
\usage{
importance(fitids = NULL, vars = NULL, result = NULL, normalize = TRUE)
}
\arguments{
\item{fitids}{Vector of fit fitids, or NULL to run for all finished fits}

\item{vars}{Vector of variables to restrict analysis to. Default = \verb{\{*\}},
all variables. \code{vars} is processed by \code{find_orthos}, and may include file names,
portable names, search names and regular expressions of file and portable names.
For example, you could use \verb{vars = 'summer | low, mid} to look at importances
only for summer season low and mid tides.}

\item{result}{File name to write results to. If NULL, one will be constructed.}

\item{normalize}{If TRUE, normalize importance by Kappa, so better fits get more importance}
}
\description{
Produce a summary of variable importance across multiple fits
}
