% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_unet.R
\name{prep_unet}
\alias{prep_unet}
\title{Pre-process data for U-Net}
\usage{
prep_unet(
  model_name,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{model_name}{The model name, which is also the name of a \code{.yml} parameter file in the \code{pars}
directory. This file must contain the following:
\itemize{
\item year: the year to fit
\item orthos: portable names of all orthophotos to include
\item patch: size in pixels
\item depth: number of of downsampling stages
\item classes: vector of target classes
\item holdout: percent of data to hold out for validation
}}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}. These take priority
over the function's defaults.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error handling. Use this
for debugging. If you get unrecovered errors, the job won't be added to the jobs database. Has
no effect if local = FALSE.}

\item{comment}{Optional slurmcollie comment}
}
\description{
Creates numpy arrays ready for fitting in U-Net. Result files are placed in \verb{<site>/unet/<model>}.
}
