% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_plot.R
\name{screen_plot}
\alias{screen_plot}
\title{Display plots for screen}
\usage{
screen_plot(extent, input, output, session)
}
\arguments{
\item{extent}{Which plot to show, one of \code{full}, \code{inset1}, or \code{inset2}}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Display plots for screen
}
\keyword{internal}
