% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_names.R
\name{search_names}
\alias{search_names}
\title{Find orthoimages that match search names}
\usage{
search_names(descrip)
}
\arguments{
\item{descrip}{Description string. See details.}
}
\value{
A named list of all orthophoto attributes designated in the description
}
\description{
Processes user-friendly generalized orthoimage search names, such as
\verb{mica, swir | ortho | low | summer | 2020:2022}, returning a list of matching category values
from \code{categories} in \code{pars.yml}. These descriptions allow selecting multiple orthoimages clearly
and simply (e.g., \code{mica | low} selects all low-tide Mica images). Importantly for cross-site
modeling, dates are replaced by seasons.
}
\details{
Categories are separated by \code{|} (spaces are optional anywhere except within a name). Use commas
to separate multiple tags in a category. Pairs of ordinal categories such as seasons and dates
may be provided separated by a colon to designate a sequence; e.g., \code{2019:2022} is the same as
\verb{2019, 2020, 2021, 2022}. Tags may include modifiers (for instance, high tide may be modified as
\code{high-spring}). Modified tags are returned as two items: first the base tag(s), such as
\code{high}, and then the tag with modification, \verb{c(high}, \verb{spring)}.

Non-existent tags generally result in an error with an informative message. Note that, for
multiple tags, the category is determined by the first tag, so \verb{sprang, summer, fall} will report
all three seasons as errors, even though the second two are correct.

An exact filename may be used as the description by prepending a \code{*}, e.g.,
\verb{*x22Sep18_OTH_Low_Mica_DEM_NoTiles.tif}.
}
\examples{
require(saltmarsh)
init()
search_names('mid-in, mid-out, high')
search_names('mica, swir, p4 | ortho | high-spring | spring:fall | 2019:2022')
search_names('mica, swir | ortho, dem | low:high | spring | 2018')
search_names('2022 | oth | mid | mica | ortho | mean-w5')
search_names('20x22 | other | muddle | micro')     # this throws an error
search_names('*x22Sep18_OTH_Low_Mica_DEM_NoTiles.tif')
}
