% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{Pull season out of salt marsh imagery files}
\usage{
seasons(files)
}
\arguments{
\item{files}{Vector of imagery file names}
}
\value{
Vector of seasons corresponding to files
}
\description{
For now, this returns months. Eventually, it will return our defined flight seasons.
}
\details{
Robust to wildly haphazard file and date formats (e.g., 18Jun2025, 18June25); does
a pretty good job of hunting down the date from the filename. Dates must be in \code{dmy}
format, followed by an underscore. For file names with two dates, finds the first.
}
\keyword{internal}
