% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_build_input_stack.R
\name{unet_build_input_stack}
\alias{unet_build_input_stack}
\title{Build input stack for a given portable ortho name}
\usage{
unet_build_input_stack(config)
}
\arguments{
\item{config}{Named \code{config} list, including
\itemize{
\item \code{config$fpath} Path to flights
\item \code{config$orthos} Vector of orthophoto names (typically a 5-band Mica, NDVI, NDRE, and a matching DEM)
\item \code{config$type} Vector of ortho type corresponding to \code{orthos}
\item \code{config$bands} Number of bands in each ortho
}}

\item{fpath}{Path to flights directory}
}
\value{
SpatRaster with all bands (typically Blue, Green, Red, NIR, RedEdge, NDVI, NDRE, DEM)
}
\description{
Build input stack for a given portable ortho name
}
