% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet_patch_stats.R
\name{unet_patch_stats}
\alias{unet_patch_stats}
\title{Summary stats for extracted patches}
\usage{
unet_patch_stats(patch_data)
}
\arguments{
\item{patch_data}{Extracted patches from \code{unet_extract_training_patches}}
}
\value{
List with train_stats, val_stats, and test_stats data frames
}
\description{
Reports stats separately for train, val, and test masks.
Hope to see single class patches < 80\%.
}
\keyword{internal}
