% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_more.R
\name{upscale_more}
\alias{upscale_more}
\title{Adds more metrics to an already existing upscaled clone}
\usage{
upscale_more(
  site,
  newsite,
  cellsize = 1,
  vars = "{*}",
  minscore = 0,
  maxmissing = 20,
  metrics = "all",
  cache = TRUE,
  resources = NULL,
  local = FALSE,
  trap = TRUE,
  comment = NULL
)
}
\arguments{
\item{site}{Site name}

\item{newsite}{Name for cloned site}

\item{cellsize}{Cell size for new site (m)}

\item{vars}{File names, portable names, regex matching either, or search names
selecting files to upscale. See Image naming in
\href{https://github.com/UMass-UAS-Salt-Marsh/marshmap/blob/main/README.md}{README}
for details. The default is \verb{\{*\}}, which will include all variables.}

\item{minscore}{Minimum score for orthos. Files with a minimum score of less than
this are excluded. Default is 0, but rejected orthos are always
excluded.}

\item{maxmissing}{Maximum percent missing in orthos. Files with percent missing greater
than this are excluded.}

\item{metrics}{A list of metrics, or 'all' for all metrics. May include any of:
\itemize{
\item 'mean' Mean (may already have been done by \code{upscale_clone})
\item \code{sd} Standard deviation
\item \code{q05}, \code{q10}, \code{q25}, \code{median}, \code{q75}, \code{q90}, and \code{q95} Quantiles
\item \code{r0595}, \code{r1090}, \code{iqr} Quantile ranges: 5th-95th, 10th-90th, and interquartile range
\item \code{skewness} and \code{kurtosis}, for Ryan
}}

\item{cache}{If TRUE, build cached images for \code{screen}}

\item{resources}{Slurm launch resources. See \link[slurmcollie]{launch}. These take priority
#'    over the function's defaults.}

\item{local}{If TRUE, run locally; otherwise, spawn a batch run on Unity}

\item{trap}{If TRUE, trap errors in local mode; if FALSE, use normal R error handling. Use this
for debugging. If you get unrecovered errors, the job won't be added to the jobs database. Has
no effect if local = FALSE.}

\item{comment}{Optional slurmcollie comment}
}
\description{
\itemize{
\item this is experimental
}
}
\details{
\strong{Note}: all metrics \emph{must} be added to pars.yml under category: derive.
}
